Spawn space crabs first iteration
Fixed numerous issues with newly spawned space crabs and the transition
to their attack state.

The first issue had to do with stubbing in a temporary instantiation for
the Ship property of SpaceCrabState, instead of assigning it to default
with a null forgiveness operator. This masked the fact that the Attack
state's ready method was not calling `base._Ready()`. It also meant that
the Ship's agent was not getting initialized properly, causing NRE's
later down the road.

Lesson learned (until furhter notice): Always call `base._Ready()` when
applicable, and use the null forgiveness operator for non-nullable
members and properties that don't need to be instantiated until _Ready()
is called.