Implement persistent space crab spawning
When space crabs are killed, a percentage is calculated based on the
number of space crabs that are alive vs the max number of allowed space
crabs. If this percentage falls below a threshold value, space crabs are
spawned up to the maximum allowed.

Fixed an issue with positioning space crabs before they are added to the
game world.